move connection logic to Database model
