+++ mathesar/models/exceptions.py
@@ -1,0 +1,6 @@
+class NoConnectionAvailable(Exception):
+    pass
+
+
+class NoAdminConnectionAvailable(Exception):
+    pass
